rename point2point to point2pointipam
