+++ pkg/networkservice/ipam/point2pointipam/server_test.go
@@ -1,4 +1,4 @@
-package point2point
+package point2pointipam
 
 import (
 	"context"
